경매 참여 api 옮기기 : product -> bid로
