+++ backend/src/main/java/com/backend/controller/auction/AuctionController.java
@@ -1,0 +1,34 @@
+package com.backend.controller.auction;
+
+import com.backend.domain.auction.BidList;
+import com.backend.service.auction.AuctionService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/api/bids")
+public class AuctionController {
+
+    private final AuctionService service;
+
+    @PostMapping("join")
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity joinProduct(@RequestBody BidList bid, Authentication authentication) {
+        //클라이언트로 부터 받은 userId(상품의 주인)와 토큰을 가지고 있는 userId가 같다면 참여 못함
+        if (service.hasAccess(bid.getProductId(), authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        } else {
+            service.upsertBidPrice(bid);
+            return ResponseEntity.ok().build();
+        }
+    }
+
+}
