+++ backend/src/main/java/com/backend/mapper/product/ProductMapper.java
@@ -1,6 +1,5 @@
 package com.backend.mapper.product;
 
-import com.backend.domain.auction.BidList;
 import com.backend.domain.chat.ChatProduct;
 import com.backend.domain.chat.ChatRoom;
 import com.backend.domain.product.Product;
@@ -145,30 +144,10 @@ public interface ProductMapper {
     @Select("SELECT COUNT(*) FROM product_like WHERE product_id=#{productId}")
     int selectCountLikeByProductId(Integer id);
 
-    @Insert("""
-            INSERT INTO bid_list (product_id, user_id, bid_price)
-            VALUES (#{productId}, #{userId}, #{bidPrice})
-            """)
-    void insertBidPrice(BidList bid);
 
     @Delete("DELETE FROM product_like WHERE product_id=#{productId}")
     int deleteLikeByProductId(Integer productId);
 
-    @Select("""
-            SELECT COUNT(*)
-            FROM bid_list
-            WHERE product_id = #{productId}
-            AND user_id = #{userId}
-            """)
-    boolean existsBid(Integer productId, Integer userId);
-
-    @Update("""
-            UPDATE bid_list
-            SET bid_price = #{bidPrice}
-            WHERE product_id = #{productId}
-            AND user_id = #{userId}
-            """)
-    int updateBidPrice(BidList bid);
 
     @Select("""
             SELECT p.id,
